requirements.txt
mlx
numpy
transformers
torch
huggingface_hub
Pillow
pytest
